# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0",
        "title": "@conjoon/php-ms-imapuser",
        "license": {
            "name": "MIT"
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/MailAccounts/{mailAccountId}/MailFolders": {
            "get": {
                "summary": "Returns all available mail boxes belonging to a specific mail account",
                "operationId": "mailFolders",
                "parameters": [
                    {
                        "$ref":"#/components/parameters/mailAccountId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Available MailFolders belonging to a specific MailAccount",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/MailFolder"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/MailAccounts/{mailAccountId}/MailFolders/{mailFolderId}/MessageItems/{messageItemId}&target=MessageBody" : {
            "get" : {
                "summary" : "Returns the MessageBody for the specified key fragments",
                "operationId": "getMessageBody",
                "parameters":[{
                    "$ref":"#/components/parameters/mailAccountId"
                },
                    {
                        "$ref":"#/components/parameters/mailFolderId"
                    },
                    {
                        "$ref":"#/components/parameters/messageItemId"
                    }],
                "responses" : {
                    "200" : {
                        "description": "The MessageBody found for the specified MailAccount, MailFolder and message id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/MessageBody"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/MailAccounts/{mailAccountId}/MailFolders/{mailFolderId}/MessageItems/{messageItemId}&target=MessageItem" : {
            "get" : {
                "summary" : "Returns the MessageItem for the specified key fragments",
                "operationId": "getMessageItem",
                "parameters":[{
                    "$ref":"#/components/parameters/mailAccountId"
                },
                    {
                        "$ref":"#/components/parameters/mailFolderId"
                    },
                    {
                        "$ref":"#/components/parameters/messageItemId"
                    }],
                "responses" : {
                    "200" : {
                        "description": "The MessageItem found for the specified MailAccount, MailFolder and message id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/MessageItem"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put" : {
                "summary" : "Updates the specific MessageItem with the submitted information.",
                "operationId": "putMessageItem",
                "parameters": [{
                    "$ref":"#/components/parameters/mailAccountId"
                },
                    {
                        "$ref":"#/components/parameters/mailFolderId"
                    },
                    {
                        "$ref":"#/components/parameters/messageItemId"
                    }, {
                        "name" : "origin",
                        "required" : false,
                        "description" : "Required to let the server know that the operation on the MessageItem is related to a recently created message, i.e. if the request is chained by creating a MessageDraft, for which now the envlope informations should be updated.",
                        "in" : "query",
                        "schema" : {
                            "type" : "string",
                            "enum" : ["create"]
                        }
                    }, {
                        "name" : "action",
                        "required" : false,
                        "description" : "Requests a move operation for the specified MessageItem. Requires additional query-parameter `mailFolderId`",
                        "in": "query",
                        "schema" : {
                            "type" : "string",
                            "enum" : ["move"]
                        }
                    }, {
                        "name" : "mailFolderId",
                        "required" : false,
                        "description" : "A new location for the MessageItem, if `action` represents a `move`-operation. Not to be confused with the path-parameter `{mailFolderId}`, which is required to identify the resource. The response must contain the MesageItem for this type of request.",
                        "in" : "query",
                        "schema" : {
                            "type" : "string"
                        }
                    }, {
                        "name": "seen",
                        "required" : false,
                        "description" : "Marks the MessageItem as seen/unseen",
                        "in" : "query",
                        "schema" : {
                            "type" : "boolean"
                        }
                    }, {
                        "name" : "flagged",
                        "required" : false,
                        "description" : "Marks the MessageItem as flagged/unflagged",
                        "in" : "query",
                        "schema" : {
                            "type" : "boolean"
                        }
                    }, {
                        "name" : "draft",
                        "required" : false,
                        "description" : "Marks the MessageItem as a draft",
                        "in" : "query",
                        "schema" : {
                            "type" : "boolean"
                        }
                    }],
                "responses":{
                    "200": {
                        "description": "The updated message information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf" : [{
                                        "$ref": "#/components/schemas/putMessageItemResponse"
                                    }, {
                                        "$ref": "#/components/schemas/moveMessageItemResponse"
                                    }]
                                }
                            }
                        }
                    }
                }

            }
        }
    },
    "components": {
        "parameters": {
            "mailAccountId" : {
                "name": "mailAccountId",
                "in": "path",
                "required": true,
                "description": "An id of a MailAccount that should be targeted for the operation",
                "schema": {
                    "type": "string"
                }
            },
            "mailFolderId" : {
                "name": "mailFolderId",
                "in": "path",
                "required": true,
                "description": "An id of the MailFolder that should be targeted for the operation",
                "schema": {
                    "type": "string"
                }
            },
            "messageItemId" : {
                "name": "messageItemId",
                "in": "path",
                "required": true,
                "description": "The id of the MessageItem that should be targeted for the operation",
                "schema": {
                    "type": "string"
                }
            }
        },
        "schemas": {
            "MailFolder": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "mailAccountId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "unreadCount": {
                        "type": "integer"
                    },
                    "folderType": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "$ref": "#/components/schemas/MailFolder"
                        }
                    }
                }
            },
...
            "MessageBody" : {
                "allOf": [{
                    "$ref": "#/components/schemas/compoundKey"
                }],
                "type": "object",
                "properties": {
                    "textHtml" : {
                        "type" : "string"
                    },
                    "textPlain" : {
                        "type" : "string"
                    }
                }
            },
...
            "MessageItem": {
                "allOf": [{
                    "$ref": "#/components/schemas/compoundKey"
                }],
                "type": "object",
                "properties": {
                    "date" : {
                        "type": "object"
                    },
                    "from": {
                        "type": "string"
                    },
                    "seen": {
                        "type": "boolean"
                    },
                    "answered" : {
                        "type": "boolean"
                    },
                    "flagged" : {
                        "type": "boolean"
                    },
                    "draft" : {
                        "type": "boolean"
                    },
                    "recent" : {
                        "type": "boolean"
                    },
                    "hasAttachments" : {
                        "type" : "boolean"
                    },
                    "to" : {
                        "type" : "array"
                    },
                    "previewText" : {
                        "type" : "string"
                    },
                    "size" : {
                        "type" : "integer"
                    },
                    "cn_deleted" : {
                        "type" : "boolean"
                    },
                    "cn_moved" : {
                        "type" : "boolean"
                    }
                }
            },
            "putMessageItemResponse" : {
                "allOf" : [{
                    "$ref" : "#/components/schemas/compoundKey"
                }],
                "type" : "object",
                "properties" : {
                    "seen" : {
                        "type": "boolean"
                    },
                    "flagged": {
                        "type": "boolean"
                    }
                }

            },
            "moveMessageItemResponse" : {
                "type":"object",
                "properties" : {
                    "success" : {
                        "type" : "boolean"
                    },
                    "data" : {
                        "$ref" : "#/components/schemas/MessageItem"
                    }
                }
            },
            "compoundKey" : {
                "type" : "object",
                "properties" : {
                    "id": {
                        "type": "string"
                    },
                    "mailAccountId": {
                        "type" : "string"
                    },
                    "mailFolderId":{
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "BasicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
    "security": [
        {
            "BasicAuth": []
        }
    ]
}